Update top level namespace for -lib-auth
