+++ tracdap-libs/tracdap-lib-auth/src/main/java/org/finos/tracdap/auth/login/LoginContent.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.finos.tracdap.common.auth.login;
+package org.finos.tracdap.auth.login;
 
 import io.netty.handler.codec.http.HttpResponseStatus;
 import org.finos.tracdap.common.exception.EUnexpected;
