+++ tracdap-libs/tracdap-lib-auth/src/main/java/org/finos/tracdap/auth/login/simple/BasicLoginProvider.java
@@ -15,9 +15,9 @@
  * limitations under the License.
  */
 
-package org.finos.tracdap.common.auth.login.simple;
+package org.finos.tracdap.auth.login.simple;
 
-import org.finos.tracdap.common.auth.login.*;
+import org.finos.tracdap.auth.login.*;
 import org.finos.tracdap.common.config.ConfigManager;
 import org.finos.tracdap.common.http.Http1Headers;
 import org.finos.tracdap.common.http.CommonHttpRequest;
