+++ tracdap-libs/tracdap-lib-auth/src/main/java/org/finos/tracdap/auth/login/simple/SqlUserDb.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.finos.tracdap.common.auth.login.simple;
+package org.finos.tracdap.auth.login.simple;
 
 import org.finos.tracdap.common.config.ConfigManager;
 import org.finos.tracdap.common.db.JdbcSetup;
