+++ tracdap-libs/tracdap-lib-auth/src/main/resources/META-INF/services/org.finos.tracdap.common.plugin.ITracPlugin
@@ -14,4 +14,4 @@
 # limitations under the License.
 
 # Include a number of common auth providers as part of the core implementation
-org.finos.tracdap.common.auth.login.simple.SimpleLoginPlugin
+org.finos.tracdap.auth.login.simple.SimpleLoginPlugin
