+++ tracdap-libs/tracdap-lib-auth/src/test/java/org/finos/tracdap/auth/login/LoginRefreshTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.finos.tracdap.common.auth.login;
+package org.finos.tracdap.auth.login;
 
 import org.finos.tracdap.common.auth.internal.*;
 import org.finos.tracdap.config.AuthenticationConfig;
