+++ tracdap-libs/tracdap-lib-common/src/main/java/org/finos/tracdap/common/plugin/PluginServiceInfo.java
@@ -70,8 +70,8 @@ public class PluginServiceInfo {
     static final Map<String, String> SERVICE_TYPES = Map.ofEntries(
             Map.entry("org.finos.tracdap.common.config.IConfigLoader", CONFIG_SERVICE_TYPE),
             Map.entry("org.finos.tracdap.common.config.ISecretLoader", SECRETS_SERVICE_TYPE),
-            Map.entry("org.finos.tracdap.common.auth.provider.IAuthProvider", AUTH_PROVIDER_SERVICE_TYPE),
-            Map.entry("org.finos.tracdap.common.auth.login.ILoginProvider", LOGIN_PROVIDER_SERVICE_TYPE),
+            Map.entry("org.finos.tracdap.auth.provider.IAuthProvider", AUTH_PROVIDER_SERVICE_TYPE),
+            Map.entry("org.finos.tracdap.auth.login.ILoginProvider", LOGIN_PROVIDER_SERVICE_TYPE),
             Map.entry("org.finos.tracdap.common.codec.ICodec", FORMAT_SERVICE_TYPE),
             Map.entry("org.finos.tracdap.common.storage.IFileStorage", FILE_STORAGE_SERVICE_TYPE),
             Map.entry("org.finos.tracdap.common.storage.IDataStorage", DATA_STORAGE_SERVICE_TYPE),
